fix(content): Don't convert value of "content"
The value of content should not be converted based on the flow direction
of the document, as it will usually not be the intended result since
`right` is not the RLT equvalent of `left` in the context of `content`.

The equvalent of `left` would be `اليسار` or `يسار` in Arabic and
`שמאל` or `שמאלה` in Hebrew.