feat(getValueDoppelganger): Ignore css property values with `var()`
Fixes #

Since css variables can be more or less anything, added an extra utility
function to match for the presence of `var()` in the css property value,
since there's really no way to know how to actually compute the RTL
doppelganger.

I was initially worried that `translateX(var())` might be an edge case
that needs to be handled, but after testing, the result is currently the
same since the regexp will never match with the `var()` being used
inside the `translateX` call.

Since the original issue is only with `boxShadow` or `textShadow` I'm
open to any suggestions that move the `containsCssVar` check only for
that property value converter.